fix: use POS Register report for dashboard data and with 'POS Register' to fetch dashboard analytics.
